requirements.txt
Django==1.11.2
django-appconf==1.0.2
django-bootstrap3==8.2.3
django-ckeditor==5.2.2
django-colorful==1.2
django-imagekit==4.0.1
olefile==0.44
pilkit==2.0
Pillow==4.1.1
pytz==2017.2
six==1.10.0
